Add RegainPermissionsFragment; rework nav logic
The current plan is that the user will be forced to grant permissions as
part of the tutorial. I think this is cleaner than having to show them
a separate page. However, if the user later goes and manually revokes a
permission, we shouldn't show them the whole tutorial again. Instead,
they are shown a new page specifically designed for this situation. This
works both on app open and if permissions are revoked while the app is
running.

Other changes:

 - The app now uses SharedPreferences to remember whether the tutorial
   has been seen, rather than just showing it if it doesn't have the
   necessary permissions.

 - Some permissions-related things used in multiple files have been
   moved to util/Permissions.kt.

 - Android KTX core extensions have been added as a dependency.