Update to androidx.appcompat 1.1.0
In AppCompat 1.1.0 (and Fragment 1.1.0, which it depends on), a lot of
things gained @Nullable or @NonNull annotations. Because of that, this
update allowed (and in some cases, required) small changes throughout.

The constructor I had been using for FragmentPagerAdapter was also
deprecated in this release, so I switched to the recommended new one.